+++ libs/base/src/data/mod.rs
@@ -3,21 +3,25 @@
 #![doc = crate::_DOC_DATA!()]
 //
 
-mod codec;
-mod iter;
-mod key;
+pub mod codec;
+pub mod iter;
+// pub mod key;
 
-crate::items! { // structural access: _mods, _all
+crate::items! { // structural access: _mods, _pub_mods, _all
     #[allow(unused)]
     pub use _mods::*;
 
     mod _mods { #![allow(unused)]
-        pub use super::codec::_all::*;
-        pub use super::iter::_all::*;
-        pub use super::key::_all::*;
+    }
+    mod _pub_mods {
+        pub use super::{
+            codec::_all::*,
+            iter::_all::*,
+            // key::_all::*,
+        };
     }
     pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]
-        pub use super::_mods::*;
+        pub use super::{_mods::*, _pub_mods::*};
     }
 }