+++ libs/base/src/work/future/mod.rs
@@ -4,6 +4,7 @@
 //
 
 mod coroutine;
+mod reexports;
 
 crate::items! { // structural access: _mods, _pub_mods, _all
     #[allow(unused)]
@@ -14,6 +15,7 @@ crate::items! { // structural access: _mods, _pub_mods, _all
     mod _mods { #![allow(unused)]
         pub use super::{
             coroutine::_all::*,
+            reexports::*,
         };
     }
     mod _pub_mods { #![allow(unused)]