+++ libs/base/src/work/sync/atomic/mod.rs
@@ -3,7 +3,7 @@
 #![doc = crate::_DOC_WORK_SYNC_ATOMIC!()]
 //
 
-pub mod reexports;
+mod reexports;
 
 crate::items! { // structural access: _mods, _pub_mods, _all, _always
     #[allow(unused)]
@@ -12,11 +12,11 @@ crate::items! { // structural access: _mods, _pub_mods, _all, _always
     pub use _pub_mods::*;
 
     mod _mods {
-    }
-    mod _pub_mods { #![allow(unused)]
         pub use super::{
             reexports::*,
         };
+    }
+    mod _pub_mods { #![allow(unused)]
     }
     pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]